+++ internal/handler/registry/mysql.go
@@ -51,6 +51,7 @@ func init() {
 			EnableNestedPushdown:    opts.EnableNestedPushdown,
 			CappedCleanupPercentage: opts.CappedCleanupPercentage,
 			CappedCleanupInterval:   opts.CappedCleanupInterval,
+			TTLCleanupInterval:      opts.TTLCleanupInterval,
 			EnableNewAuth:           opts.EnableNewAuth,
 			BatchSize:               opts.BatchSize,
 			MaxBsonObjectSizeBytes:  opts.MaxBsonObjectSizeBytes,